+++ mayan/apps/document_states/tests/test_workflow_instance_models.py
@@ -128,6 +128,19 @@ class WorkflowInstanceModelTestCase(
         self.assertEqual(events[0].target, self._test_document)
         self.assertEqual(events[0].verb, event_document_created.id)
 
+
+class WorkflowInstanceTransitionModelTestCase(
+    WorkflowInstanceTestMixin, GenericDocumentTestCase
+):
+    auto_upload_test_document = False
+
+    def setUp(self):
+        super().setUp()
+        self._create_test_workflow_template(add_test_document_type=True)
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition()
+
     def test_workflow_template_transition_no_condition(self):
         self._clear_events()